refactor: migrate price fields from double to decimal for precision
This commit updates various price-related fields across the application to use the `decimal` type instead of `double` to ensure better precision in financial calculations. Additionally, it includes a new migration to reflect these changes in the database schema. The refactor affects DTOs, entities, and related services to maintain consistency throughout the codebase.

+++ Backend/Application/Reports/Dtos/CategoryPerformanceReportResponse.cs
@@ -5,7 +5,7 @@ public class CategoryPerformanceReportResponse
     public Guid CategoryId { get; set; }
     public string CategoryName { get; set; } = null!;
     public int ProductCount { get; set; }
-    public double TotalRevenue { get; set; }
+    public decimal TotalRevenue { get; set; }
     public int TotalUnitsSold { get; set; }
     public double AverageRating { get; set; }
 }
